+++ src/core/types.ts
@@ -1,5 +1,0 @@
-export enum NamingStrategy {
-    MIXED = 0,
-    USE_FILE_SEQUENCE = 1,
-    USE_FILE_PATH = 2,
-}